refactor: separated logic into independent files, and moved all lib files into web/app/lib (#33)
