+++ web/app/lib/consts.js
@@ -1,0 +1,3 @@
+export const primaryColor = '#25AE88'
+export const warnColor = '#f1c40f'
+export const errorColor = '#e74c3c'
