+++ web/app/lib/device.js
